feat: implement transaction status retrieval and booking update logic
